requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
courlan==1.3.2
dateparser==1.2.2
discord.py==2.6.0
frozenlist==1.7.0
htmldate==1.9.3
idna==3.10
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
multidict==6.6.4
propcache==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.5
six==1.17.0
soupsieve==2.7
tld==0.13.1
trafilatura==2.0.0
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
Flask==3.1.0
